- Updated ConvertIHolonToICelestialBody & ConvertIHolonToICelestialSpace methods in Mapper in OASIS.API.Core.
- Updated ICelestialBody interface (made some properties long rather than int) in OASIS.API.Core.

- Updated ICelestialHolon interface  (made some properties long rather than int) in OASIS.API.Core.

- Fixed bugs in LoadChildHolonsRecursive method overloads in OASIS.API.Core.

- Added System namespace to the objectSerializer in Init method to fix issues in the MongoDBOASIS Provider.

- Updated Asteroid, AsteroidModel, CelestialBodyAbstract & Comet in the SQLDBOASIS Provider to reflect changes made above.

- Added COSMICVersion & STARODKVersion properties to OASISBootLoader and also it now shows the current versions after the oasis has booted along with the OASIS Runtime version.

- Changed some CelestialProperties such as Mass, Weight, Density, RotationPeriod, OrbitPeriod, GravitaionalPull, DistanceFromParentStarInMetres, RotationSpeed, etc to long rather than int in CelestialBody in STAR ODK.

- Marked LoadAsync & Load methods Obsolete in CelestialBody in STAR ODK. Recommend to use LoadAsync<T> and Load<T> instead.

- Updated LoadAsync<T> & Load<T> in CelestialBody in STAR ODK.

- Re-wrote SaveAsync, Save, Save<T> & SaveAsync<T> in CelestialBody in STAR ODK.

- Added new SetParentIds, SetMetaData, ProcessSaveResult & ProcessSaveResult<T> methods to CelestialBody in STAR ODK.

- Updated Save, SaveAsync<T> & Save<T> methods in CelestialSpace in STAR ODK.

- Added new SetCelestialHolonMetaData method to CelestialSpace in STAR ODK.

- Changed Size & Age from int to long CelestialHolon in STAR ODK.

- Done lots more upgrades and improvements to the STAR ODK Test Harness/CLI, it now has a cool new asci tree allowing users to visualaize the objects in COSMIC ORM such as holons, zomes, celestialbodies, celestialspaces etc.

- Updated Test method in STAR ODK Test Harness/CLI.

- Updated LoadCelestialBodyAsync method in STAR ODK Test Harness/CLI.

- Updated LoadHolonAsync method in STAR ODK Test Harness/CLI.

- Updated ShowHolonProperties method in STAR ODK Test Harness/CLI.

- Updated ShowZomesAndHolons method in STAR ODK Test Harness/CLI.

- Updated ShowHolons method in STAR ODK Test Harness/CLI.

- Added new ShowHolonBasicProperties method to STAR ODK Test Harness/CLI.

+++ NextGenSoftware.OASIS.API.Core/Interfaces/STAR/CelestialBodies/ICelestialBody.cs
@@ -12,20 +12,20 @@ namespace NextGenSoftware.OASIS.API.Core.Interfaces.STAR
     {
         ICelestialBodyCore CelestialBodyCore { get; set; }
         int CurrentOrbitAngleOfParentStar { get; set; }
-        int Density { get; set; }
-        int DistanceFromParentStarInMetres { get; set; }
-        int GravitaionalPull { get; set; }
+        long Density { get; set; }
+        long DistanceFromParentStarInMetres { get; set; }
+        long GravitaionalPull { get; set; }
         Guid Id { get; set; }
-        int Mass { get; set; }
+        long Mass { get; set; }
         int NumberActiveAvatars { get; set; }
         int NumberRegisteredAvatars { get; set; }
-        int OrbitPeriod { get; set; }
+        long OrbitPeriod { get; set; }
         int OrbitPositionFromParentStar { get; set; }
         Dictionary<ProviderType, string> ProviderUniqueStorageKey { get; set; }
-        int RotationPeriod { get; set; }
-        int RotationSpeed { get; set; }
+        long RotationPeriod { get; set; }
+        long RotationSpeed { get; set; }
         int TiltAngle { get; set; }
-        int Weight { get; set; }
+        long Weight { get; set; }
 
         event EventDelegates.CelestialBodyError OnCelestialBodyError;
         event EventDelegates.CelestialBodyLoaded OnCelestialBodyLoaded;
